+++ cli/src/commands/git/clone.rs
@@ -283,6 +283,7 @@ fn configure_remote(
         tx.repo_mut(),
         remote_name,
         source,
+        None,
         fetch_tags.as_fetch_tags(),
         bookmark_expr,
     )?;