+++ lib/tests/test_git.rs
@@ -3761,6 +3761,7 @@ fn test_fetch_with_fetch_tags_override() {
         tx.repo_mut(),
         "origin".as_ref(),
         &source_git_repo.path().display().to_string(),
+        None,
         gix::remote::fetch::Tags::None,
         &StringExpression::all(),
     )
@@ -3801,6 +3802,7 @@ fn test_fetch_with_fetch_tags_override() {
         tx.repo_mut(),
         "originAllTags".as_ref(),
         &source_git_repo.path().display().to_string(),
+        None,
         gix::remote::fetch::Tags::All,
         &StringExpression::all(),
     )
@@ -4869,6 +4871,7 @@ fn test_remote_remove_refs() {
         tx.repo_mut(),
         "foo".as_ref(),
         "https://example.com/",
+        None,
         Default::default(),
         &StringExpression::all(),
     )
@@ -4919,6 +4922,7 @@ fn test_remote_rename_refs() {
         tx.repo_mut(),
         "foo".as_ref(),
         "https://example.com/",
+        None,
         Default::default(),
         &StringExpression::all(),
     )
@@ -4996,6 +5000,7 @@ fn test_remote_add_with_tags_specification() {
             tx.repo_mut(),
             remote_name.as_ref(),
             "https://example.com/",
+            None,
             fetch_tags,
             &StringExpression::all(),
         )
@@ -5032,6 +5037,7 @@ fn test_remote_add_with_refspecs() {
         tx.repo_mut(),
         "origin".as_ref(),
         "https://example.com/",
+        None,
         gix::remote::fetch::Tags::default(),
         &bookmark_expr,
     )
@@ -5123,8 +5129,9 @@ fn test_set_remote_urls() {
         tx.repo_mut(),
         remote_name.as_ref(),
         "https://example.com/repo/path",
-        gix::remote::fetch::Tags::None,
         None,
+        gix::remote::fetch::Tags::None,
+        &StringExpression::all(),
     )
     .unwrap();